# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Smart Pill Doser POC.",
        "title": "Smart Pill Doser API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/caregivers": {
            "get": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Caregivers"
                ],
                "summary": "List caregivers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/api.CaregiverRequest"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Caregivers"
                ],
                "summary": "Add caregiver",
                "parameters": [
                    {
                        "description": "Caregiver details",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CaregiverRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.CaregiverRequest"
                        }
                    }
                }
            }
        },
        "/caregivers/{id}": {
            "delete": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Caregivers"
                ],
                "summary": "Delete caregiver",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Caregiver ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeletedIDResponse"
                        }
                    }
                }
            }
        },
        "/device/heartbeat": {
            "post": {
                "security": [
                    {
                        "HardwareAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Hardware"
                ],
                "summary": "Device heartbeat",
                "parameters": [
                    {
                        "description": "Battery Level",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.HealthResponse"
                        }
                    }
                }
            }
        },
        "/intake-logs": {
            "post": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "tags": [
                    "IntakeLogs"
                ],
                "summary": "Log medication intake",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/schedules": {
            "get": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "List schedules",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Schedule"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Create schedule",
                "parameters": [
                    {
                        "description": "Schedule times",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Schedule"
                        }
                    }
                }
            }
        },
        "/schedules/{id}": {
            "delete": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Schedules"
                ],
                "summary": "Delete schedule",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Schedule ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.DeletedIDResponse"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Create user",
                "parameters": [
                    {
                        "description": "User email",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/api.CreateUserResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "MobileAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Users"
                ],
                "summary": "Patch user notification settings",
                "parameters": [
                    {
                        "description": "Notification settings",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.PatchUserRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        }
    },
    "definitions": {
        "api.CaregiverRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "api.CreateScheduleRequest": {
            "type": "object",
            "properties": {
                "time1": {
                    "description": "e.g. '08:00:00'",
                    "type": "string"
                },
                "time2": {
                    "description": "e.g. '20:00:00'",
                    "type": "string"
                }
            }
        },
        "api.CreateUserRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "api.CreateUserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.DeletedIDResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "api.HealthResponse": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
        "api.PatchUserRequest": {
            "type": "object",
            "properties": {
                "notify_after_minutes": {
                    "type": "integer"
                },
                "notify_caregivers_after_retries": {
                    "type": "integer"
                }
            }
        },
        "models.Schedule": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "time1": {
                    "description": "e.g. '08:00:00'",
                    "type": "string"
                },
                "time2": {
                    "description": "e.g. '20:00:00'",
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "HardwareAuth": {
            "type": "apiKey",
            "name": "X-Device-Serial",
            "in": "header"
        },
        "MobileAuth": {
            "type": "apiKey",
            "name": "X-User-ID",
            "in": "header"
        }
    }
}